requirements.txt
groq
openai
PyPDF2
pdfplumber
langchain
arxiv
fastapi
uvicorn
python-dotenv
